// src/core/models/Cover.ts
export type BuildStatus =
    | { Error: null }
    | { Building: null }
    | { Success: null }
    | { Pending: null };
export type CanisterType = { Rust: null } | { Motoko: null };

export type Error = { BuildInProgress: null };
// ...
export interface Verification {
  updated_at: string;
  updated_by: Principal;
  canister_id: Principal;
  dfx_version: string;
  build_status: BuildStatus;
  canister_name: string;
  commit_hash: string;
  canister_type: [] | [CanisterType];
  repo_url: string;
  rust_version: [] | [string];
  optimize_count: number;
  build_url: [] | [string];
  wasm_hash: [] | [string];
}